add eac metric
